# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Description here.",
      "version": "1.0.0",
      "title": "Swagger Netflix",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "tags": [
      {
        "name": "user",
        "description": "Operations about user",
        "externalDocs": {
          "description": "Find out more about our store",
          "url": "http://swagger.io"
        }
      }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
      "/account/login": {
        "post": {
          "tags": [
            "account"
          ],
          "summary": "Logs user into the system",
          "description": "",
          "operationId": "loginUser",
          "produces": [
              "application/json",
              "application/xml"
          ],
          "parameters": [
            {
                "in": "body",
                "name": "body",
                "description": "order placed for purchasing the pet",
                "required": true,
                "schema": {
                  "$ref": "#/definitions/login"
                }
              }
           
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "string"
              },
              "headers": {
                "X-Rate-Limit": {
                  "type": "integer",
                  "format": "int32",
                  "description": "calls per hour allowed by the user"
                },
                "X-Expires-After": {
                  "type": "string",
                  "format": "date-time",
                  "description": "date in UTC when token expires"
                }
              }
            },
            "400": {
              "description": "Invalid username/password supplied"
            }
          }
        }
      },
      "/account/create": {
        "post": {
          "tags": [
            "account"
          ],
          "summary": "Logs user into the system",
          "description": "",
          "operationId": "createUser",
          "produces": [
              "application/json",
              "application/xml"
          ],
          "parameters": [
            {
                "in": "body",
                "name": "body",
                "description": "order placed for purchasing the pet",
                "required": true,
                "schema": {
                  "$ref": "#/definitions/account"
                }
              }
           
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "string"
              },
              "headers": {
                "X-Rate-Limit": {
                  "type": "integer",
                  "format": "int32",
                  "description": "calls per hour allowed by the user"
                },
                "X-Expires-After": {
                  "type": "string",
                  "format": "date-time",
                  "description": "date in UTC when token expires"
                }
              }
            },
            "400": {
              "description": "Invalid username/password supplied"
            }
          }
        }
      },
      "/account/upload": {
        "post": {
          "tags": [
            "account"
          ],
          "summary": "Logs user into the system",
          "description": "",
          "operationId": "upload-video",
          "produces": [
              "application/json",
              "application/xml"
          ],
          "parameters": [
            {
                "in": "body",
                "name": "body",
                "description": "order placed for purchasing the pet",
                "required": true,
                "schema": {
                  "$ref": "#/definitions/account"
                }
              }
           
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "string"
              },
              "headers": {
                "X-Rate-Limit": {
                  "type": "integer",
                  "format": "int32",
                  "description": "calls per hour allowed by the user"
                },
                "X-Expires-After": {
                  "type": "string",
                  "format": "date-time",
                  "description": "date in UTC when token expires"
                }
              }
            },
            "400": {
              "description": "Invalid username/password supplied"
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
      "/user/add":{
          "post":{
            "tags":["user"],
            "summary":"Creates user",
            "description":"",
            "operationId":"createNewUser",
            "produces":["application/json","application/xml"],
            "parameters":[
              {
                "in":"body",
                "name":"body",
                "description":"",
                "required":true,
                "schema":{"$ref":"#/definitions/createUser"}
              }
            ],
            "responses":{
              "200":{
                "description":"successful operation",
                "schema":{
                  "$ref":"#/definitions/ApiResponse"
                }
              }
            }
          }
      },
      "/user/all":{
        "get":{
          "tags":[
            "user"
          ],
          "produces":[
            "application/json"
          ],
          "responses":{

          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      }
    },
    "securityDefinitions": {
      "api_key": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    },
    "definitions": {
        "account":{
            "type":"object",
            "properties":{
                "firstName":{"type":"string"},
                "lastName":{"type":"string"},
                "email":{"type":"string"},
                "password":{"type":"string"}
            }
        },
        "login":{
            "type":"object",
            "properties":{
                "email":{"type":"string"},
                "password":{"type":"string"}
            }
        },
        "createUser":{
          "type":"object",
          "properties":{
            "firstName":{"type":"string"},
            "lastName":{"type":"string"},
            "phone":{"type":"string"},
            "email":{"type":"string"},
            "password":{"type":"string"},
            "avatar":{"type":"string"}
        }
        },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      }
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    }
  }